Simple tests are added. Some fixes.

+++ app/routes/pairs.py
@@ -1,8 +1,10 @@
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Depends
 from models import Pair, Dancer, PairResponse
 from schemas import StatusType
 from db.session import SessionDep
 from sqlmodel import select
+from auth_handler import get_current_user
+from fastapi import status
 
 app = APIRouter(prefix='/pairs', tags=['pairs'])
 
@@ -54,7 +56,9 @@ def read_pair(pair_id: int, session: SessionDep) -> PairResponse:
     )
 
 @app.delete("/{pair_id}")
-def delete_pair(pair_id: int, session: SessionDep):
+def delete_pair(pair_id: int, 
+                session: SessionDep,
+                current_user: dict = Depends(get_current_user)):
     """
     Удалить пару и обновить статусы танцоров.
     
@@ -76,37 +80,48 @@ def delete_pair(pair_id: int, session: SessionDep):
     pair = session.get(Pair, pair_id)
     if not pair:
         raise HTTPException(status_code=404, detail="Pair not found")
+    
+    if current_user.user_type == "DANCER": 
+        if not(current_user.dancer_id is None) and \
+            current_user.dancer_id != pair.dancer1_id and \
+            current_user.dancer_id != pair.dancer2_id:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="No enough right to delete pair. " \
+                "User should be on of participant of pair or admin." 
+            )
+    else:
 
-    # Получаем связанных танцоров
-    dancer1 = session.get(Dancer, pair.dancer1_id)
-    dancer2 = session.get(Dancer, pair.dancer2_id)
+        # Получаем связанных танцоров
+        dancer1 = session.get(Dancer, pair.dancer1_id)
+        dancer2 = session.get(Dancer, pair.dancer2_id)
 
-    # Удаляем пару
-    session.delete(pair)
+        # Удаляем пару
+        session.delete(pair)
 
-    # Проверяем другие существующие пары
-    dancer1_pairs = session.exec(
-        select(Pair).where(
-            (Pair.dancer1_id == dancer1.id) |
-            (Pair.dancer2_id == dancer1.id)
-        )
-    ).all()
+        # Проверяем другие существующие пары
+        dancer1_pairs = session.exec(
+            select(Pair).where(
+                (Pair.dancer1_id == dancer1.id) |
+                (Pair.dancer2_id == dancer1.id)
+            )
+        ).all()
 
-    dancer2_pairs = session.exec(
-        select(Pair).where(
-            (Pair.dancer1_id == dancer2.id) |
-            (Pair.dancer2_id == dancer2.id)
-        )
-    ).all()
+        dancer2_pairs = session.exec(
+            select(Pair).where(
+                (Pair.dancer1_id == dancer2.id) |
+                (Pair.dancer2_id == dancer2.id)
+            )
+        ).all()
 
-    # Обновляем статусы если нужно
-    if not dancer1_pairs and dancer1.status == StatusType.IN_PAIR:
-        dancer1.status = StatusType.IN_SEARCH
-        session.add(dancer1)
+        # Обновляем статусы если нужно
+        if not dancer1_pairs and dancer1.status == StatusType.IN_PAIR:
+            dancer1.status = StatusType.IN_SEARCH
+            session.add(dancer1)
 
-    if not dancer2_pairs and dancer2.status == StatusType.IN_PAIR:
-        dancer2.status = StatusType.IN_SEARCH
-        session.add(dancer2)
+        if not dancer2_pairs and dancer2.status == StatusType.IN_PAIR:
+            dancer2.status = StatusType.IN_SEARCH
+            session.add(dancer2)
 
-    session.commit()
-    return {"ok": True}
+        session.commit()
+        return {"ok": True}
